add: added item detail edit

+++ app/src/main/java/com/project/suitcase/view/ui/fragment/HomeFragment.kt
@@ -2,14 +2,18 @@ package com.project.suitcase.view.ui.fragment
 
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.project.suitcase.R
 import com.project.suitcase.databinding.FragmentHomeBinding
 import com.project.suitcase.view.adapter.ParentTripAdapter
+import com.project.suitcase.view.ui.activity.ItemDetailActivity
 import com.project.suitcase.view.ui.activity.ItemListActivity
 import com.project.suitcase.view.viewmodel.DeleteAllTripViewModelEvent
 import com.project.suitcase.view.viewmodel.HomeFragmentUiState
@@ -47,7 +51,18 @@ class HomeFragment : Fragment() {
             adapter = parentTripAdapter
         }
         binding?.btnDeleteAllTrips?.setOnClickListener {
-            homeFragmentViewModel.deleteAllTrip()
+
+            MaterialAlertDialogBuilder(requireContext(),
+                R.style.ThemeOverlay_App_MaterialAlertDialog)
+                .setTitle("Delete all Trip?")
+                .setMessage("All trips and their items will be removed from the list.")
+                .setNegativeButton("NO") { dialog, which ->
+                    //nothing
+                }
+                .setPositiveButton("YES") { dialog, which ->
+                    homeFragmentViewModel.deleteAllTrip()
+                }
+                .show()
         }
 
         parentTripAdapter?.onItemClick = {
@@ -55,6 +70,14 @@ class HomeFragment : Fragment() {
                 putExtra("tripId", it.tripId)
                 putExtra("tripName", it.tripName)
                 putExtra("tripDate", it.date)
+            }
+            startActivity(intent)
+        }
+        parentTripAdapter?.onChildItemClick = {
+            val intent = Intent(requireContext(), ItemDetailActivity::class.java).apply {
+                putExtra("itemID", it.itemId)
+                putExtra("tripID", it.tripId)
+                putExtra("itemImage", it.itemImage)
             }
             startActivity(intent)
         }
@@ -77,6 +100,7 @@ class HomeFragment : Fragment() {
             when(event) {
                 is HomeFragmentViewModelEvent.Error -> {
                     Toast.makeText(requireContext(), event.error, Toast.LENGTH_SHORT).show()
+                    Log.e("HomeFragment", event.error)
                 }
                 is HomeFragmentViewModelEvent.Success -> {
                     binding?.rvTrip?.visibility = View.VISIBLE
